# management/registry/internal/handler/subscription_group.go
package handler

import (
	"encoding/json"
	"fmt"
	"net/http"
	"strconv"
	"strings"
	"time"

	"hive/registry/internal/middleware"
	"hive/registry/internal/model"

	"gorm.io/gorm"
)

type CreateGroupRequest struct {
	Name string `json:"name"`
}

type SetGroupNodesRequest struct {
	Nodes []string `json:"nodes"`
}

type ResetTokenResponse struct {
	Token string `json:"token"`
}

// buildGroupClashYAML generates a group-specific Clash YAML.
// overrideUUID, when non-empty, forces all proxies to use that UUID (per-subscription).
func (h *Handler) buildGroupClashYAML(groupName string, nodes []model.Node, overrideUUID string) string {
	return buildFullClashYAML(groupName, nodes, h.Config.XrayPath, overrideUUID)
}

// HandleListGroups handles GET /admin/subscription-groups
// @Summary      获取订阅组列表
// @ID           AdminListSubscriptionGroups
// @Description  返回所有订阅组
// @Tags         admin
// @Produce      json
// @Success      200 {array} model.SubscriptionGroup
// @Failure      500 {object} ErrorResponse
// @Router       /admin/subscription-groups [get]
func (h *Handler) HandleListGroups(w http.ResponseWriter, r *http.Request) {
	groups := make([]model.SubscriptionGroup, 0)
	if err := h.DB.Raw("SELECT g.id, g.name, g.token, COUNT(gn.node_mac) AS node_count, g.created_at, g.updated_at FROM subscription_groups g LEFT JOIN subscription_group_nodes gn ON gn.group_id = g.id GROUP BY g.id ORDER BY g.id").Scan(&groups).Error; err != nil {
		h.jsonErr(w, http.StatusInternalServerError, "db: "+err.Error())
		return
	}
	h.jsonOK(w, groups)
}

// HandleCreateGroup handles POST /admin/subscription-groups
// @Summary      创建订阅组
// @ID           AdminCreateSubscriptionGroup
// @Description  创建新的订阅组
// @Tags         admin
// @Accept       json
// @Produce      json
// @Param        body body CreateGroupRequest true "订阅组信息"
// @Success      200 {object} model.SubscriptionGroup
// @Failure      400 {object} ErrorResponse
// @Failure      500 {object} ErrorResponse
// @Router       /admin/subscription-groups [post]
func (h *Handler) HandleCreateGroup(w http.ResponseWriter, r *http.Request) {
	var req CreateGroupRequest
	if err := json.NewDecoder(r.Body).Decode(&req); err != nil {
		h.jsonErr(w, http.StatusBadRequest, "invalid JSON: "+err.Error())
		return
	}
	req.Name = strings.TrimSpace(req.Name)
	if req.Name == "" {
		h.jsonErr(w, http.StatusBadRequest, "name is required")
		return
	}

	token, err := generateToken()
	if err != nil {
		h.jsonErr(w, http.StatusInternalServerError, "token: "+err.Error())
		return
	}

	now := time.Now().UTC().Format(model.TimeLayout)
	var id uint
	if err := h.DB.Transaction(func(tx *gorm.DB) error {
		if err := tx.Exec(
			"INSERT INTO subscription_groups (name, token, created_at, updated_at) VALUES (?, ?, ?, ?)",
			req.Name, token, now, now,
		).Error; err != nil {
			return err
		}
		return tx.Raw("SELECT LAST_INSERT_ID()").Scan(&id).Error
	}); err != nil {
		h.jsonErr(w, http.StatusInternalServerError, "db: "+err.Error())
		return
	}
	h.jsonOK(w, model.SubscriptionGroup{
		ID:        id,
		Name:      req.Name,
		Token:     token,
		NodeCount: 0,
		CreatedAt: now,
		UpdatedAt: now,
	})
}

// HandleDeleteGroup handles DELETE /admin/subscription-groups/{id}
// @Summary      删除订阅组
// @ID           AdminDeleteSubscriptionGroup
// @Description  根据 ID 删除订阅组
// @Tags         admin
// @Produce      json
// @Param        id path int true "订阅组 ID"
// @Success      200 {object} StatusResponse
// @Failure      400 {object} ErrorResponse
// @Failure      404 {object} ErrorResponse
// @Failure      500 {object} ErrorResponse
// @Router       /admin/subscription-groups/{id} [delete]
func (h *Handler) HandleDeleteGroup(w http.ResponseWriter, r *http.Request) {
	id, err := strconv.ParseUint(r.PathValue("id"), 10, 64)
	if err != nil {
		h.jsonErr(w, http.StatusBadRequest, "invalid id")
		return
	}
	var name string
	if err := h.DB.Raw("SELECT name FROM subscription_groups WHERE id=?", id).Scan(&name).Error; err != nil || name == "" {
		h.jsonErr(w, http.StatusNotFound, "group not found")
		return
	}
	if err := h.DB.Exec("DELETE FROM subscription_groups WHERE id=?", id).Error; err != nil {
		h.jsonErr(w, http.StatusInternalServerError, "db: "+err.Error())
		return
	}
	h.jsonOK(w, map[string]string{"status": "ok"})
}

// HandleGetGroupNodes handles GET /admin/subscription-groups/{id}/nodes
// @Summary      获取订阅组节点列表
// @ID           AdminGetSubscriptionGroupNodes
// @Description  返回指定订阅组的节点 MAC 列表
// @Tags         admin
// @Produce      json
// @Param        id path int true "订阅组 ID"
// @Success      200 {array} string
// @Failure      400 {object} ErrorResponse
// @Failure      404 {object} ErrorResponse
// @Failure      500 {object} ErrorResponse
// @Router       /admin/subscription-groups/{id}/nodes [get]
func (h *Handler) HandleGetGroupNodes(w http.ResponseWriter, r *http.Request) {
	id, err := strconv.ParseUint(r.PathValue("id"), 10, 64)
	if err != nil {
		h.jsonErr(w, http.StatusBadRequest, "invalid id")
		return
	}
	var cnt int64
	if err := h.DB.Table("subscription_groups").Where("id=?", id).Count(&cnt).Error; err != nil || cnt == 0 {
		h.jsonErr(w, http.StatusNotFound, "group not found")
		return
	}
	var macs []string
	if err := h.DB.Model(&model.SubscriptionGroupNode{}).Where("group_id=?", id).Order("node_mac").Pluck("node_mac", &macs).Error; err != nil {
		h.jsonErr(w, http.StatusInternalServerError, "db: "+err.Error())
		return
	}
	if macs == nil {
		macs = make([]string, 0)
	}
	h.jsonOK(w, macs)
}

// HandleSetGroupNodes handles PUT /admin/subscription-groups/{id}/nodes
// @Summary      设置订阅组节点
// @ID           AdminSetSubscriptionGroupNodes
// @Description  替换指定订阅组的节点列表
// @Tags         admin
// @Accept       json
// @Produce      json
// @Param        id path int true "订阅组 ID"
// @Param        body body SetGroupNodesRequest true "节点列表"
// @Success      200 {object} StatusResponse
// @Failure      400 {object} ErrorResponse
// @Failure      404 {object} ErrorResponse
// @Failure      500 {object} ErrorResponse
// @Router       /admin/subscription-groups/{id}/nodes [put]
func (h *Handler) HandleSetGroupNodes(w http.ResponseWriter, r *http.Request) {
	id, err := strconv.ParseUint(r.PathValue("id"), 10, 64)
	if err != nil {
		h.jsonErr(w, http.StatusBadRequest, "invalid id")
		return
	}
	var cnt int64
	if err := h.DB.Table("subscription_groups").Where("id=?", id).Count(&cnt).Error; err != nil || cnt == 0 {
		h.jsonErr(w, http.StatusNotFound, "group not found")
		return
	}

	var req SetGroupNodesRequest
	if err := json.NewDecoder(r.Body).Decode(&req); err != nil {
		h.jsonErr(w, http.StatusBadRequest, "invalid JSON: "+err.Error())
		return
	}
	if req.Nodes == nil {
		req.Nodes = []string{}
	}

	if err := h.DB.Transaction(func(tx *gorm.DB) error {
		if err := tx.Where("group_id=?", id).Delete(&model.SubscriptionGroupNode{}).Error; err != nil {
			return err
		}
		for _, mac := range req.Nodes {
			tx.Exec("INSERT IGNORE INTO subscription_group_nodes (group_id, node_mac) VALUES (?, ?)", id, mac)
		}
		now := time.Now().UTC().Format(model.TimeLayout)
		tx.Exec("UPDATE subscription_groups SET updated_at=? WHERE id=?", now, id)
		return nil
	}); err != nil {
		h.jsonErr(w, http.StatusInternalServerError, "db: "+err.Error())
		return
	}
	h.jsonOK(w, map[string]string{"status": "ok"})
}

// HandleResetGroupToken handles POST /admin/subscription-groups/{id}/reset-token
// @Summary      重置订阅组 Token
// @ID           AdminResetSubscriptionGroupToken
// @Description  为指定订阅组生成新的访问 Token
// @Tags         admin
// @Produce      json
// @Param        id path int true "订阅组 ID"
// @Success      200 {object} ResetTokenResponse
// @Failure      400 {object} ErrorResponse
// @Failure      404 {object} ErrorResponse
// @Failure      500 {object} ErrorResponse
// @Router       /admin/subscription-groups/{id}/reset-token [post]
func (h *Handler) HandleResetGroupToken(w http.ResponseWriter, r *http.Request) {
	id, err := strconv.ParseUint(r.PathValue("id"), 10, 64)
	if err != nil {
		h.jsonErr(w, http.StatusBadRequest, "invalid id")
		return
	}
	var cnt int64
	if err := h.DB.Table("subscription_groups").Where("id=?", id).Count(&cnt).Error; err != nil || cnt == 0 {
		h.jsonErr(w, http.StatusNotFound, "group not found")
		return
	}

	token, err := generateToken()
	if err != nil {
		h.jsonErr(w, http.StatusInternalServerError, "token: "+err.Error())
		return
	}
	now := time.Now().UTC().Format(model.TimeLayout)
	if err := h.DB.Exec("UPDATE subscription_groups SET token=?, updated_at=? WHERE id=?", token, now, id).Error; err != nil {
		h.jsonErr(w, http.StatusInternalServerError, "db: "+err.Error())
		return
	}
	h.jsonOK(w, ResetTokenResponse{Token: token})
}

// HandlePublicGroupClash handles GET /s/{token}
// @Summary      公开订阅组 Clash 配置
// @ID           PublicGroupClash
// @Description  通过 Token 获取订阅组的 Clash YAML 配置（无需认证）
// @Tags         subscription
// @Produce      plain
// @Param        token path string true "订阅组 Token"
// @Success      200 {string} string "Clash subscription"
// @Failure      404 {string} string "Not Found"
// @Failure      500 {string} string "Internal Server Error"
// @Router       /s/{token} [get]
func (h *Handler) HandlePublicGroupClash(w http.ResponseWriter, r *http.Request) {
	middleware.SubscriptionRequestsTotal.WithLabelValues("public_group_clash").Inc()

	token := r.PathValue("token")
	if len(token) != 64 {
		http.NotFound(w, r)
		return
	}

	var sg struct {
		ID   uint
		Name string
	}
	if err := h.DB.Raw("SELECT id, name FROM subscription_groups WHERE token=?", token).Scan(&sg).Error; err != nil || sg.ID == 0 {
		http.NotFound(w, r)
		return
	}

	var macs []string
	if err := h.DB.Model(&model.SubscriptionGroupNode{}).Where("group_id=?", sg.ID).Pluck("node_mac", &macs).Error; err != nil {
		http.Error(w, "db error", http.StatusInternalServerError)
		return
	}

	nodes, err := h.queryNodesByMACs(macs)
	if err != nil {
		http.Error(w, "db error", http.StatusInternalServerError)
		return
	}

	yaml := h.buildGroupClashYAML(sg.Name, nodes, "")
	w.Header().Set("Content-Type", "text/plain; charset=utf-8")
	w.Header().Set("Content-Disposition", fmt.Sprintf(`attachment; filename="hive-%s.yaml"`, sg.Name))
	fmt.Fprint(w, yaml)
}