# management/registry/openapi/openapi.yaml
basePath: /
definitions:
...
  handler.CreateGroupRequest:
    properties:
      name:
        type: string
    type: object
...
  handler.ErrorResponse:
    properties:
      error:
        example: error message
        type: string
    type: object
...
  handler.ResetTokenResponse:
    properties:
      token:
        type: string
    type: object
...
  handler.SetGroupNodesRequest:
    properties:
      nodes:
        items:
          type: string
        type: array
    type: object
...
  handler.StatusResponse:
    properties:
      status:
        example: ok
        type: string
    type: object
...
  model.SubscriptionGroup:
    properties:
      created_at:
        type: string
      id:
        type: integer
      name:
        type: string
      node_count:
        type: integer
      token:
        type: string
      updated_at:
        type: string
    type: object
...
info:
  contact: {}
  description: Hive management plane API for nodes and subscriptions.
  title: Hive Node Registry API
  version: 0.1.0
paths:
...
  /admin/subscription-groups:
    get:
      description: 返回所有订阅组
      operationId: AdminListSubscriptionGroups
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/model.SubscriptionGroup'
            type: array
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
      summary: 获取订阅组列表
      tags:
      - admin
    post:
      consumes:
      - application/json
      description: 创建新的订阅组
      operationId: AdminCreateSubscriptionGroup
      parameters:
      - description: 订阅组信息
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/handler.CreateGroupRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.SubscriptionGroup'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
      summary: 创建订阅组
      tags:
      - admin
  /admin/subscription-groups/{id}:
    delete:
      description: 根据 ID 删除订阅组
      operationId: AdminDeleteSubscriptionGroup
      parameters:
      - description: 订阅组 ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handler.StatusResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
      summary: 删除订阅组
      tags:
      - admin
  /admin/subscription-groups/{id}/nodes:
    get:
      description: 返回指定订阅组的节点 MAC 列表
      operationId: AdminGetSubscriptionGroupNodes
      parameters:
      - description: 订阅组 ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              type: string
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
      summary: 获取订阅组节点列表
      tags:
      - admin
    put:
      consumes:
      - application/json
      description: 替换指定订阅组的节点列表
      operationId: AdminSetSubscriptionGroupNodes
      parameters:
      - description: 订阅组 ID
        in: path
        name: id
        required: true
        type: integer
      - description: 节点列表
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/handler.SetGroupNodesRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handler.StatusResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
      summary: 设置订阅组节点
      tags:
      - admin
  /admin/subscription-groups/{id}/reset-token:
    post:
      description: 为指定订阅组生成新的访问 Token
      operationId: AdminResetSubscriptionGroupToken
      parameters:
      - description: 订阅组 ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handler.ResetTokenResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
      summary: 重置订阅组 Token
      tags:
      - admin
...
  /s/{token}:
    get:
      description: 通过 Token 获取订阅组的 Clash YAML 配置（无需认证）
      operationId: PublicGroupClash
      parameters:
      - description: 订阅组 Token
        in: path
        name: token
        required: true
        type: string
      produces:
      - text/plain
      responses:
        "200":
          description: Clash subscription
          schema:
            type: string
        "404":
          description: Not Found
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      summary: 公开订阅组 Clash 配置
      tags:
      - subscription
  /subscription:
    get:
      description: 返回 base64 编码的 VLESS 订阅链接
      operationId: SubscriptionVless
      produces:
      - text/plain
      responses:
        "200":
          description: VLESS subscription
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
      summary: 获取 VLESS 订阅
      tags:
      - subscription
  /subscription/clash:
    get:
      description: 返回 Clash/Mihomo YAML 格式订阅配置
      operationId: SubscriptionClash
      produces:
      - text/plain
      responses:
        "200":
          description: Clash subscription
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
      summary: 获取 Clash 订阅
      tags:
      - subscription
schemes:
- https
- http
securityDefinitions:
  AdminSessionCookie:
    description: '"HttpOnly session cookie for admin UI"'
    in: cookie
    name: hive_admin_session
    type: apiKey
  BearerAuth:
    description: '"Authorization: Bearer <API_SECRET>"'
    in: header
    name: Authorization
    type: apiKey
  CustomerSessionCookie:
    description: '"HttpOnly session cookie for customer portal"'
    in: cookie
    name: hive_customer_session
    type: apiKey
swagger: "2.0"